Types moved to separate files, namespaces adjusted
