+++ src/CaptainHook.Domain/RequestValidators/AddSubscriberRequestValidator.cs
@@ -1,4 +1,5 @@
 ﻿using CaptainHook.Domain.Requests;
+using CaptainHook.Domain.Requests.Subscribers;
 using FluentValidation;
 
 namespace CaptainHook.Domain.RequestValidators
